+++ Symusic/src/java/it/fivano/symusic/core/parser/GenericParser.java
@@ -147,7 +147,7 @@ public abstract class GenericParser {
 		String pass = doc.getElementsByAttributeValue("name", "pass").get(0).attr("value");
 //		System.out.println(jschl_vc);
 		Elements scriptElements = doc.getElementsByTag("script");
-		Integer numberCalcLine = null;
+		Long numberCalcLine = null;
 		for (Element element :scriptElements ){
 			for (DataNode node : element.dataNodes()) {
 				String text = node.getWholeData();
@@ -169,7 +169,7 @@ public abstract class GenericParser {
 //			System.out.println(numberCalcLine);
 		}
 
-		int jschl_answer = 0;
+		long jschl_answer = 0;
 		if(numberCalcLine!=null) {
 
 			jschl_answer = numberCalcLine;
